requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
pytest==7.4.3
python-dotenv==1.0.0
apache-airflow==2.7.3
selenium==4.16.0
webdriver-manager==4.0.1
pydantic==2.5.2
jsonschema